pyproject.toml
[project]
name = "obj-viewer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pillow>=11.3.0",
    "pyglm>=2.8.2",
    "scipy>=1.16.0",
    "slangpy>=0.31.0",
    "trimesh>=4.7.1",
]

[dependency-groups]
dev = [
    "ipython>=9.4.0",
]
